Revert "Merge commit 'a14720e2e97d6151dd8fe154e2c04dd656f727a4' into ProductDtail-UI"
This reverts commit ac062a26b1d01c4ac37c8344d2aaa8983425a41d, reversing
changes made to d7e01aa77aadabbfda3a8ef4ae6f6f6666639529.

+++ src/utils/apiHandler.js
@@ -1,12 +1,11 @@
-import { data } from "autoprefixer";
 import axios, { AxiosError } from "axios";
 
 const baseUrl = "https://api.storefront.wdb.skooldio.dev";
 
 export const getData = async (url, options = {}) => {
   try {
-    const { data } = await axios.get(`${baseUrl}/${url}`, options);
-    return data;
+    const res = await axios.get(`${baseUrl}/${url}`, options);
+    return res;
   } catch (err) {
     if (err instanceof AxiosError) {
       console.log("axios error");
@@ -24,15 +23,14 @@ export const getData = async (url, options = {}) => {
 // in validate body return valid body or Error
 // thus checking body before fire http req
 export const postData = async (url, body, options = {}) => {
-  console.log(body);
   if (body instanceof Error) {
     console.log("body validation error");
     console.error(body);
     return;
   }
   try {
-    const { data } = await axios.post(`${baseUrl}/${url}`, body, options);
-    return data;
+    const res = await axios.post(`${baseUrl}/${url}`, body, options);
+    return res;
   } catch (err) {
     if (err instanceof AxiosError) {
       console.log("axios error");
@@ -48,10 +46,14 @@ export const postData = async (url, body, options = {}) => {
 };
 
 export const deleteData = async (url, options = {}) => {
+  if (body instanceof Error) {
+    console.log("body validation error");
+    console.error(body);
+    return;
+  }
   try {
     const res = await axios.delete(`${baseUrl}/${url}`, options);
-    console.info(res);
-    return res.data;
+    return res;
   } catch (err) {
     if (err instanceof AxiosError) {
       console.log("axios error");
@@ -73,8 +75,8 @@ export const updateData = async (url, body, options = {}) => {
     return;
   }
   try {
-    const { data } = await axios.patch(`${baseUrl}/${url}`, body, options);
-    return data;
+    const res = await axios.patch(`${baseUrl}/${url}`, body, options);
+    return res;
   } catch (err) {
     if (err instanceof AxiosError) {
       console.log("axios error");
